Move MainContainer and NotFound to component folder ...
