+++ src/component/signup/Signup.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import MainContainer from '../../MainContainer';
+import MainContainer from '../maincontainer/MainContainer';
 import { FormErrors } from './FormErrors';
 import { Redirect, withRouter } from 'react-router-dom';
 
@@ -22,7 +22,7 @@ class Signup extends Component {
         }
     }
 
-    // update state variable whenever user inputs. 
+    // update state variable whenever user inputs.
     onChange = (e) => {
         this.handleUserInput(e);
     }
@@ -71,8 +71,8 @@ class Signup extends Component {
     }
 
     onSubmit = (e) => {
-        e.preventDefault();   //prevents actual submission. 
-        console.log(this.state); // just checking values; remove once done. 
+        e.preventDefault();   //prevents actual submission.
+        console.log(this.state); // just checking values; remove once done.
         this.setState(() => ({ toProfile: true }));
         /* later on implement database entry */
     }
@@ -141,4 +141,4 @@ class Signup extends Component {
     }
 }
 
-export default Signup; 
+export default Signup;
